+++ test/Well.Shift.t.sol
@@ -5,7 +5,7 @@ import {TestHelper, Balances, ConstantProduct2, IERC20} from "test/TestHelper.so
 import {IWell} from "src/interfaces/IWell.sol";
 
 contract WellShiftTest is TestHelper {
-    event Shift(uint[] reserves, IERC20 toToken, uint minAmountOut, address recipient);
+    event Shift(uint256[] reserves, IERC20 toToken, uint256 minAmountOut, address recipient);
 
     ConstantProduct2 cp;
 
@@ -15,7 +15,7 @@ contract WellShiftTest is TestHelper {
     }
 
     /// @dev Shift excess token0 into token1.
-    function testFuzz_shift(uint amount) public prank(user) {
+    function testFuzz_shift(uint256 amount) public prank(user) {
         amount = bound(amount, 1, 1000e18);
 
         // Transfer `amount` of token0 to the Well
@@ -33,16 +33,16 @@ contract WellShiftTest is TestHelper {
         assertEq(userBalanceBeforeShift.tokens[1], 0, "User should start with 0 of token1");
 
         well.sync();
-        uint minAmountOut = well.getShiftOut(tokens[1]);
-        uint[] memory calcReservesAfter = new uint[](2);
+        uint256 minAmountOut = well.getShiftOut(tokens[1]);
+        uint256[] memory calcReservesAfter = new uint256[](2);
         calcReservesAfter[0] = well.getReserves()[0];
         calcReservesAfter[1] = well.getReserves()[1] - minAmountOut;
 
         vm.expectEmit(true, true, true, true);
         emit Shift(calcReservesAfter, tokens[1], minAmountOut, _user);
-        uint amtOut = well.shift(tokens[1], minAmountOut, _user);
+        uint256 amtOut = well.shift(tokens[1], minAmountOut, _user);
 
-        uint[] memory reserves = well.getReserves();
+        uint256[] memory reserves = well.getReserves();
         Balances memory userBalanceAfterShift = getBalances(_user, well);
         Balances memory wellBalanceAfterShift = getBalances(address(well), well);
 
@@ -70,7 +70,7 @@ contract WellShiftTest is TestHelper {
     }
 
     /// @dev Shift excess token0 into token0 (just transfers the excess token0 to the user).
-    function testFuzz_shift_tokenOut(uint amount) public prank(user) {
+    function testFuzz_shift_tokenOut(uint256 amount) public prank(user) {
         amount = bound(amount, 1, 1000e18);
 
         // Transfer `amount` of token0 to the Well
@@ -87,8 +87,8 @@ contract WellShiftTest is TestHelper {
         assertEq(userBalanceBeforeShift.tokens[1], 0, "User should start with 0 of token1");
 
         well.sync();
-        uint minAmountOut = well.getShiftOut(tokens[0]);
-        uint[] memory calcReservesAfter = new uint[](2);
+        uint256 minAmountOut = well.getShiftOut(tokens[0]);
+        uint256[] memory calcReservesAfter = new uint256[](2);
         calcReservesAfter[0] = well.getReserves()[0] - minAmountOut;
         calcReservesAfter[1] = well.getReserves()[1];
 
@@ -97,7 +97,7 @@ contract WellShiftTest is TestHelper {
         // Shift the imbalanced token as the token out
         well.shift(tokens[0], 0, _user);
 
-        uint[] memory reserves = well.getReserves();
+        uint256[] memory reserves = well.getReserves();
         Balances memory userBalanceAfterShift = getBalances(_user, well);
         Balances memory wellBalanceAfterShift = getBalances(address(well), well);
 
@@ -134,7 +134,7 @@ contract WellShiftTest is TestHelper {
 
         well.shift(tokens[1], 0, _user);
 
-        uint[] memory reserves = well.getReserves();
+        uint256[] memory reserves = well.getReserves();
         Balances memory userBalanceAfterShift = getBalances(_user, well);
         Balances memory wellBalanceAfterShift = getBalances(address(well), well);
 
@@ -148,7 +148,7 @@ contract WellShiftTest is TestHelper {
         checkInvariant(address(well));
     }
 
-    function test_shift_fail_slippage(uint amount) public prank(user) {
+    function test_shift_fail_slippage(uint256 amount) public prank(user) {
         amount = bound(amount, 1, 1000e18);
 
         // Transfer `amount` of token0 to the Well
@@ -157,8 +157,8 @@ contract WellShiftTest is TestHelper {
         assertEq(wellBalanceBeforeShift.tokens[0], 1000e18 + amount, "Well should have received token0");
         assertEq(wellBalanceBeforeShift.tokens[1], 1000e18, "Well should have NOT have received token1");
 
-        uint amountOut = well.getShiftOut(tokens[1]);
-        vm.expectRevert(abi.encodeWithSelector(IWell.SlippageOut.selector, amountOut, type(uint).max));
-        well.shift(tokens[1], type(uint).max, user);
+        uint256 amountOut = well.getShiftOut(tokens[1]);
+        vm.expectRevert(abi.encodeWithSelector(IWell.SlippageOut.selector, amountOut, type(uint256).max));
+        well.shift(tokens[1], type(uint256).max, user);
     }
 }
